requirements.txt
-i https://pypi.org/simple
asgiref==3.11.1; python_version >= '3.9'
certifi==2026.1.4; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.4; python_version >= '3.7'
cryptography==46.0.5; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==6.0.2; python_version >= '3.12'
django-filter==25.2; python_version >= '3.10'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
djoser==2.3.3; python_version >= '3.9' and python_version < '4.0'
idna==3.11; python_version >= '3.8'
oauthlib==3.3.1; python_version >= '3.8'
pycparser==3.0; python_version >= '3.10'
pyjwt[crypto]==2.11.0; python_version >= '3.9'
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.5; python_version >= '3.9'
requests-oauthlib==2.0.0; python_version >= '3.4'
social-auth-app-django==5.7.0; python_version >= '3.10'
social-auth-core==4.8.5; python_version >= '3.10'
sqlparse==0.5.5; python_version >= '3.8'
urllib3==2.6.3; python_version >= '3.9'
